+++ apps/mis-server/src/services/admin.ts
@@ -14,10 +14,10 @@ import { plugin } from "@ddadaal/tsgrpc-server";
 import { ServiceError } from "@grpc/grpc-js";
 import { Status } from "@grpc/grpc-js/build/src/constants";
 import { AdminServiceServer, AdminServiceService } from "@scow/protos/build/server/admin";
+import { updateBlockStatusInSlurm } from "src/bl/block";
 import { importUsers, ImportUsersData } from "src/bl/importUsers";
 import { Account } from "src/entities/Account";
 import { StorageQuota } from "src/entities/StorageQuota";
-import { User } from "src/entities/User";
 import { parseClusterUsers } from "src/utils/slurm";
 
 export const adminServiceServer = plugin((server) => {
@@ -79,7 +79,7 @@ export const adminServiceServer = plugin((server) => {
     },
 
     importUsers: async ({ request, em, logger }) => {
-      const { data, whitelist } = request;
+      const { data, tenantName, whitelist } = request;
 
       if (!data) {
         throw <ServiceError> {
@@ -102,7 +102,7 @@ export const adminServiceServer = plugin((server) => {
         };
       }
 
-      const reply = await importUsers(data as ImportUsersData, em, whitelist, logger);
+      const reply = await importUsers(data as ImportUsersData, em, whitelist, tenantName, logger);
 
       return [reply];
 
@@ -127,20 +127,9 @@ export const adminServiceServer = plugin((server) => {
       includedAccounts.forEach((account) => {
         const a = result.accounts.find((x) => x.accountName === account.accountName)!;
         a.included = true;
-        a.owner = "该账户已导入";
       });
 
-      const includedUsers = await em.find(User, {
-        userId: { $in: result.users.map((x) => x.userId) },
-      });
-      includedUsers.forEach((user) => {
-        // https://slurm.schedmd.com/sacctmgr.html#OPT_user
-        // slurm's username defaults to lower case
-        const u = result.users.find((x) => x.userId.toLowerCase() === user.userId.toLowerCase())!;
-        u.included = true;
-        u.userName = user.name;
-      });
-
+      result.accounts.sort((a, b) => a.included === b.included ? 0 : (a.included === false ? -1 : 1));
       return [result];
     },
 
@@ -168,6 +157,11 @@ export const adminServiceServer = plugin((server) => {
       const reply = await server.ext.fetch.fetch();
 
       return [reply];
+    },
+
+    updateBlockStatus: async ({ em, logger }) => {
+      await updateBlockStatusInSlurm(em, server.ext.clusters, logger);
+      return [{}];
     },
 
   });