+++ apps/mis-web/src/apis/api.ts
@@ -32,6 +32,7 @@ import type { SetPlatformRoleSchema } from "src/pages/api/admin/setPlatformRole"
 import type { SetTenantRoleSchema } from "src/pages/api/admin/setTenantRole";
 import type { UnsetPlatformRoleSchema } from "src/pages/api/admin/unsetPlatformRole";
 import type { UnsetTenantRoleSchema } from "src/pages/api/admin/unsetTenantRole";
+import type { UpdateBlockStatusSchema } from "src/pages/api/admin/updateBlockStatus";
 import type { AuthCallbackSchema } from "src/pages/api/auth/callback";
 import type { LogoutSchema } from "src/pages/api/auth/logout";
 import type { ValidateTokenSchema } from "src/pages/api/auth/validateToken";
@@ -100,6 +101,7 @@ export const api = {
   setTenantRole: fromApi<SetTenantRoleSchema>("PUT", join(basePath, "/api/admin/setTenantRole")),
   unsetPlatformRole: fromApi<UnsetPlatformRoleSchema>("PUT", join(basePath, "/api/admin/unsetPlatformRole")),
   unsetTenantRole: fromApi<UnsetTenantRoleSchema>("PUT", join(basePath, "/api/admin/unsetTenantRole")),
+  updateBlockStatus: fromApi<UpdateBlockStatusSchema>("PUT", join(basePath, "/api/admin/updateBlockStatus")),
   authCallback: fromApi<AuthCallbackSchema>("GET", join(basePath, "/api/auth/callback")),
   logout: fromApi<LogoutSchema>("DELETE", join(basePath, "/api/auth/logout")),
   validateToken: fromApi<ValidateTokenSchema>("GET", join(basePath, "/api/auth/validateToken")),