+++ apps/mis-web/src/pageComponents/init/InitImportUsersTable.tsx
@@ -10,9 +10,7 @@
  * See the Mulan PSL v2 for more details.
  */
 
-import { FormLayout } from "@scow/lib-web/build/layouts/FormLayout";
 import { Alert, Typography } from "antd";
-import { Centered } from "src/components/layouts";
 import styled from "styled-components";
 
 import { ImportUsersTable } from "../admin/ImportUsersTable";
@@ -26,24 +24,22 @@ const AlertContainer = styled.div`
 export const InitImportUsersTable: React.FC = () => {
 
   return (
-    <Centered>
-      <FormLayout maxWidth={800}>
-        <Typography.Paragraph>您可以在此导入已有用户。 查看
-          <a target="_blank" href="https://pkuhpc.github.io/SCOW/docs/mis/business/users" rel="noreferrer">
+    <div>
+      <Typography.Paragraph>您可以在此导入已有用户。 查看
+        <a target="_blank" href="https://pkuhpc.github.io/SCOW/docs/mis/business/users" rel="noreferrer">
           此文档
-          </a>
+        </a>
           了解系统用户模型以及如何导入用户信息。
-        </Typography.Paragraph>
-        <AlertContainer>
-          <Alert
-            type="warning"
-            showIcon
-            message="如果您使用SCOW管理多个集群，SCOW系统要求多个集群具有完全相同的用户账户信息，您只需要从一个集群中导入已有用户信息即可。"
-          />
-        </AlertContainer>
-        <ImportUsersTable />
-      </FormLayout>
-    </Centered>
+      </Typography.Paragraph>
+      <AlertContainer>
+        <Alert
+          type="warning"
+          showIcon
+          message="如果您使用SCOW管理多个集群，SCOW系统要求多个集群具有完全相同的用户账户信息，您只需要从一个集群中导入已有用户信息即可。"
+        />
+      </AlertContainer>
+      <ImportUsersTable />
+    </div>
   );
 };