+++ apps/auth/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @scow/auth
 
+## 0.2.0
+
+### Minor Changes
+
+- dc9852988: 修改 ldap.searchBase 为必填
+- 84fcc4bf3: 增加配置日志输出选项功能
+- 401a21ebe: 限制登录系统可回调域名和增加相关配置
+
+### Patch Changes
+
+- Updated dependencies [6814c3427]
+- Updated dependencies [c24e21662]
+  - @scow/config@0.1.1
+  - @scow/lib-config@0.1.1
+
 ## 0.1.2
 
 ### Patch Changes