+++ apps/portal-web/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @scow/portal-web
 
+## 0.2.0
+
+### Minor Changes
+
+- 686054ac3: 增加 ListAvailableWms API
+- c2a1dff41: 自定义 footer 和 portal 的 dashboard 文本支持 HTML 标签
+- 5cd477d37: 门户系统修改 getAppAttributes API 为 getAppMetadata，从 portal-server 获取应用名称
+- 5e4f6ac58: 支持动态设置 base path
+- a9b64169c: 更新 LOGO、favicon、仪表盘图片的自定义方式
+
+### Patch Changes
+
+- 7d7bc0834: 修复 BASE_PATH 为/时，点击导航栏的 shell 跳转地址错误的问题
+- b6347a367: 更新 next.js 版本并开启自动更新
+- 9244b1079: 删除多余的 console.log
+- 93eb54ea6: 修复图片没有正确加上 base path
+- 6814c3427: 交互式应用的自定义表单可以配置提示信息等
+- Updated dependencies [c2a1dff41]
+- Updated dependencies [93eb54ea6]
+- Updated dependencies [5e4f6ac58]
+- Updated dependencies [a9b64169c]
+- Updated dependencies [6814c3427]
+- Updated dependencies [c24e21662]
+  - @scow/lib-web@0.2.0
+  - @scow/config@0.1.1
+  - @scow/lib-config@0.1.1
+  - @scow/lib-decimal@0.1.1
+  - @scow/utils@0.1.1
+
 ## 0.1.2
 
 ## 0.1.1